+++ lib/src/pxtone/pxtnUnit.cpp
@@ -10,7 +10,7 @@ pxtnUnit::pxtnUnit()
 	_bPlayed   = true;
 	_bOperated = true;
 	strcpy( _name_buf, "no name" );
-	_name_size = strlen( _name_buf );
+	_name_size = (int32_t)strlen( _name_buf );
 }
 
 pxtnUnit::~pxtnUnit()
